src/requirements.txt
akasha-terminal

python-dotenv
PyYAML

pip
setuptools
pip-autoremove